[REFACTOR] Changes Object Mother class name to singular
